@@ -32,12 +32,12 @@ publicsuffix2 = "^2.20191221"
 filetype = "^1.0.10"
 numpy = "^1.22.3"
 w3lib = "^1.22.0"
-Sphinx = { version = "^4.4.0", optional = true }
+Sphinx = { version = "^4.5.0", optional = true }
 cchardet = "^2.1.7"
 
 [tool.poetry.dev-dependencies]
-mypy = "^0.941"
-ipython = "^8.1.1"
+mypy = "^0.942"
+ipython = "^8.2.0"
 jupyterlab = "^3.3.2"
 pytest-cov = "^3.0.0"
 
